services/validation_worker/requirements.txt
# RabbitMQ client
pika==1.3.2

# Redis client
redis==5.0.1

# PostgreSQL client
psycopg2-binary==2.9.9

# Prometheus metrics
prometheus-client==0.19.0

# Environment variables
python-dotenv==1.0.0

# Additional utilities
typing-extensions==4.9.0
